OSIS-1210 Move get_warnings out of LearningContainerYear model (-> business)
